refactor: promote hub to a top-level transport layer
Now that hub.rs holds only the Hub's client, credentials, repo identity and repo
lifecycle, its home under memory/ is wrong: four layers call it, and only two of ~20 call
sites are inside the memory layer. commands/update reaches it for the self-update release
bucket, traces/source for parquet trace-repo downloads, main for the add/curate flows —
none of which involve a memory. `funes::memory::hub::hf_token()` on the path of a release
download was the tell.

src/hub.rs is now a sibling of memory, and the crate doc records it as the transport
layer. The split in the previous commit is what made this a pure move: hub.rs has no
crate:: dependencies of its own.

Also repairs the doc link that split left dangling — memory/dataset.rs pointed at
`super::hub::Memory::open`, but Memory now lives in the domain root.
