+++ src/commands/curate.rs
@@ -18,7 +18,7 @@
 //! Human- and agent-editable. A decision flipped to `exclude` later does not retract what
 //! already shipped — the remote is append-only; curation prevents, it does not undo.
 
-use crate::memory::hub;
+use crate::hub;
 use crate::memory::{self, Memory};
 use crate::memory::{dataset, hf_dataset};
 use crate::traces::jsonl;
