+++ src/commands/push.rs
@@ -21,10 +21,10 @@
 //! memory** and takes everything, as ever.
 
 use super::curate;
+use crate::hub;
 use crate::memory::card::{self, CardAction, CardCtx};
 use crate::memory::dataset;
 use crate::memory::hf_dataset::{self, Appended, Reindexed};
-use crate::memory::hub;
 use crate::memory::{self, Memory};
 use crate::{chunk, scan};
 use anyhow::{bail, Context, Result};
