+++ tests/curate_gating.rs
@@ -86,7 +86,7 @@ async fn a_project_memory_ships_only_included_sessions() {
     // A throwaway memory: unique repo name so concurrent/repeated runs don't collide.
     let nanos = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_nanos();
     let name = format!("funes-test-curate-{}-{nanos}", std::process::id());
-    if let Err(e) = funes::memory::hub::create_dataset_repo(OWNER, &name).await {
+    if let Err(e) = funes::hub::create_dataset_repo(OWNER, &name).await {
         eprintln!("skip: cannot create a scratch repo under {OWNER}: {e}");
         return;
     }
